Refactor per PR review
